chore(gatsby): Migrate query-watcher to ts (#27324)
* chore(gatsby): Migrate query-watcher to ts

* fix exports and some types

* modify import from index

* optional chaining instead of check if it is undefined
